[SD-226] Remove MonadIO from Base constraints
Problem: in previous version of this library it was possible to use AVL+
inside of STM monad. It is no longer possible due to MonadIO constraint
in Base type synonym.

Cause: Base is used in most of AVL operations but, in fact, it is
actually used only inside of `clean` function.

Fix: move MonadIO constraint from Base into clean function.